refactor(app): extract Message enum and osd_info into submodules
Move the Message enum to app/message.rs and the osd_info_for helper
to app/osd_info.rs, reducing app.rs by ~120 lines.

No behavior changes — pure structural refactor.
